feat(webshell): webshell component transformation
